Fix CI

+++ run_conan.py
@@ -46,14 +46,16 @@ def main():
         cc = os.getenv("CC")
         if cc is not None:
             new_cc = os.path.join(conan_home, "gcc")
-            os.symlink(cc, new_cc)
-            os.environ["CC"] = new_cc
+            if not os.path.exists(new_cc):
+                os.symlink(cc, new_cc)
+                os.environ["CC"] = new_cc
 
         cxx = os.getenv("CXX")
         if cxx is not None:
             new_cxx = os.path.join(conan_home, "g++")
-            os.symlink(cc, new_cxx)
-            os.environ["CC"] = new_cxx
+            if not os.path.exists(new_cxx):
+                os.symlink(cxx, new_cxx)
+                os.environ["CXX"] = new_cxx
 
     run_conan_profile_detect(conan)
     run_conan_install(conan)
